refactor search_box to reuse it in Utils module
