+++ lib/oli_web/components/delivery/utils.ex
@@ -182,6 +182,39 @@ defmodule OliWeb.Components.Delivery.Utils do
     """
   end
 
+  attr :search_term, :string, default: ""
+  attr :on_search, :string, default: "search"
+  attr :on_change, :string, default: "search"
+  attr :on_clear_search, :string, default: "clear_search"
+  attr :rest, :global, include: ~w(class)
+
+  def search_box(assigns) do
+    ~H"""
+    <form class={["flex flex-row", @rest[:class]]} phx-submit={@on_search}>
+      <div class="flex-1 relative">
+        <i class="fa-solid fa-search absolute left-4 top-4 text-gray-400 pointer-events-none text-lg">
+        </i>
+        <input
+          type="text"
+          name="search_term"
+          value={@search_term}
+          class="w-full border border-gray-400 dark:border-gray-700 rounded-lg px-12 py-3"
+          phx-change={@on_change}
+          phx-debounce="500"
+        />
+        <button
+          :if={@search_term != ""}
+          type="button"
+          class="absolute right-0 top-0 bottom-0 py-3 px-4"
+          phx-click={@on_clear_search}
+        >
+          <i class="fa-solid fa-xmark text-lg"></i>
+        </button>
+      </div>
+    </form>
+    """
+  end
+
   attr(:current_user, User)
 
   def user_icon(%{current_user: _} = assigns) do
